updated repo with crud, changed HandleValadSubmit, added total price

+++ server-prog-blazer-app/server-prog-blazer-app/models/Todo.cs
@@ -8,6 +8,6 @@ namespace server_prog_blazer_app.models
         public string ?UserEmailId { get; set; }
         public string ?TodoTitle { get; set; }
         public int StackSize { get; set; }
-        public float Price { get; set; }
+        public int Price { get; set; }
     }
 }
